+++ htmlgen/__init__.py
@@ -96,3 +96,4 @@ from .table import (
     Column,
 )
 from .time import Time
+from .video import Preload, Video